Adjusted integer types in reckonCheckDigit( )

+++ test/retail/books/ISBN13Test.java
@@ -12,7 +12,7 @@ class ISBN13Test {
     private static final long MAX_ISBN_13_NO_CHECK_DIGIT = 979987654321L;
 
     private static byte reckonCheckDigit(long digits) {
-        int sum = 0;
+        long sum = 0;
         long curr = digits;
         for (int i = 12; i > 0; i--) {
             long digit = curr % 10;
@@ -21,7 +21,7 @@ class ISBN13Test {
             sum += weighted;
             curr /= 10;
         }
-        int remainder = sum % 10;
+        long remainder = sum % 10;
         if (remainder == 0) {
             return 0;
         }
